refactor (dartpad-preview) extract bottomPanelTabs
