+++ pkgs/dartpad_preview/dartpad_frontend/lib/features/bottom_panel/views/bottom_panel.dart
@@ -6,17 +6,21 @@ import 'package:dartpad_editor/dartpad_editor.dart';
 import 'package:jaspr/dom.dart';
 import 'package:jaspr/jaspr.dart';
 
-import '../../../app_styles.dart';
+import 'bottom_panel_tabs.dart';
 import 'problems_panel.dart';
 
-/// The bottom panel showing a "Problems" tab with diagnostics.
-class BottomPanel extends StatelessComponent {
+/// The available tabs in the bottom panel.
+enum BottomPanelTab {
+  /// The problems tab showing diagnostics.
+  problems,
+}
+
+/// The bottom panel showing tabs with associated content panes.
+class BottomPanel extends StatefulComponent {
   const BottomPanel({
     required this.diagnostics,
     required this.activeFile,
     required this.onOpenDiagnostic,
-    required this.showProblems,
-    required this.onShowProblems,
     super.key,
   });
 
@@ -29,28 +33,31 @@ class BottomPanel extends StatelessComponent {
   /// Called when the user clicks a diagnostic row.
   final void Function(String fileName, Diagnostic diagnostic) onOpenDiagnostic;
 
-  /// Whether the problems panel content is visible.
-  final bool showProblems;
+  @override
+  State<BottomPanel> createState() => _BottomPanelState();
 
-  /// Shows the problems panel.
-  final void Function() onShowProblems;
+  @css
+  static List<StyleRule> get styles => _BottomPanelState.styles;
+}
+
+class _BottomPanelState extends State<BottomPanel> {
+  BottomPanelTab _activeTab = BottomPanelTab.problems;
+
+  void _selectTab(BottomPanelTab tab) {
+    setState(() {
+      _activeTab = tab;
+    });
+  }
 
   @override
   Component build(BuildContext context) {
     return div(classes: 'bottom-panel', [
-      _buildTabs(),
-      if (showProblems) _buildContent(),
-    ]);
-  }
-
-  Component _buildTabs() {
-    return div(classes: 'bottom-panel-tabs', [
-      _BottomPanelTabButton(
-        label: 'Problems',
-        countLabel: diagnostics.length.toString(),
-        active: showProblems,
-        onClick: onShowProblems,
+      BottomPanelTabs(
+        problemsCount: component.diagnostics.length,
+        activeTab: _activeTab,
+        onSelectTab: _selectTab,
       ),
+      _buildContent(),
     ]);
   }
 
@@ -58,11 +65,13 @@ class BottomPanel extends StatelessComponent {
     return div(
       classes: 'bottom-panel-content',
       [
-        ProblemsPanel(
-          diagnostics: diagnostics,
-          activeFile: activeFile,
-          onOpenDiagnostic: onOpenDiagnostic,
-        ),
+        switch (_activeTab) {
+          BottomPanelTab.problems => ProblemsPanel(
+            diagnostics: component.diagnostics,
+            activeFile: component.activeFile,
+            onOpenDiagnostic: component.onOpenDiagnostic,
+          ),
+        },
       ],
     );
   }
@@ -73,14 +82,6 @@ class BottomPanel extends StatelessComponent {
       display: .flex,
       height: 100.percent,
       flexDirection: .column,
-      flex: const .shrink(0),
-    ),
-    css('.bottom-panel .bottom-panel-tabs').styles(
-      display: .flex,
-      border: .only(
-        bottom: .solid(color: colorBorder, width: 1.px),
-      ),
-      alignItems: .stretch,
       flex: const .shrink(0),
     ),
     css('.bottom-panel .bottom-panel-content').styles(
@@ -96,74 +97,5 @@ class BottomPanel extends StatelessComponent {
       margin: .zero,
       flex: const Flex(grow: 1, basis: .zero),
     ),
-    css('.bottom-panel .bottom-panel-tab').styles(
-      display: .inlineFlex,
-      padding: .symmetric(vertical: 7.px, horizontal: 14.px),
-      border: .none,
-      outline: const Outline(style: .none),
-      cursor: .pointer,
-      userSelect: .none,
-      transition: .combine([
-        Transition('background', duration: 150.ms, curve: .ease),
-        Transition('color', duration: 150.ms, curve: .ease),
-        Transition('border-color', duration: 150.ms, curve: .ease),
-      ]),
-      alignItems: .center,
-      gap: Gap.all(6.px),
-      color: colorOnSurfaceVariant,
-      fontSize: 12.px,
-      fontWeight: .w500,
-      backgroundColor: Colors.transparent,
-    ),
-    css('.bottom-panel .bottom-panel-tab:hover').styles(
-      color: colorOnSurface,
-      backgroundColor: colorOnSurface.withOpacity(0.06),
-    ),
-    css('.bottom-panel .bottom-panel-tab.active').styles(
-      color: colorOnSurface,
-      backgroundColor: colorContainer,
-    ),
-    css('.bottom-panel .bottom-panel-tab-count').styles(
-      display: .inlineFlex,
-      padding: .symmetric(vertical: 1.px, horizontal: 4.px),
-      radius: .circular(8.px),
-      justifyContent: .center,
-      alignItems: .center,
-      color: colorOnPrimary,
-      fontSize: 10.px,
-      fontWeight: .w600,
-      backgroundColor: colorPrimary,
-    ),
-    css('.bottom-panel .bottom-panel-tabs-spacer').styles(
-      flex: const Flex(grow: 1),
-    ),
   ];
-}
-
-class _BottomPanelTabButton extends StatelessComponent {
-  const _BottomPanelTabButton({
-    required this.label,
-    this.countLabel,
-    required this.active,
-    this.onClick,
-  });
-
-  final String label;
-  final String? countLabel;
-  final bool active;
-  final void Function()? onClick;
-
-  @override
-  Component build(BuildContext context) {
-    final classes = active ? 'bottom-panel-tab active' : 'bottom-panel-tab';
-
-    return button(
-      classes: classes,
-      onClick: onClick,
-      [
-        span(classes: 'bottom-panel-tab-label', [.text(label)]),
-        if (countLabel case final countLabel?) span(classes: 'bottom-panel-tab-count', [.text(countLabel)]),
-      ],
-    );
-  }
 }
